fix(oxlint): satisfy source length gate
