+++ tests/tooling/github-workflows-release-ref-env.test.ts
@@ -1,0 +1,15 @@
+import { test } from "node:test";
+
+import {
+  assertReleasePlatformCommandsUseRefEnv,
+  readRepoFile,
+} from "./support/github-workflows.ts";
+
+test("release workflow passes release refs to shell commands through env", () => {
+  assertReleasePlatformCommandsUseRefEnv(readRepoFile(".github", "workflows", "release.yml"), [
+    "plan-release-platforms",
+    "smoke-release-packages",
+    "release-npm-native",
+    "release-npm-oxlint-plugin",
+  ]);
+});
